+++ NEWS.md
@@ -6,6 +6,7 @@ Unversioned
 
   * Comparison operators for constructing constraints (e.g. ``2x >= 1``) have been deprecated. Instead, construct the constraints explicitly in
     the ``@addConstraint`` macro to add them to the model, or in the ``@LinearConstraint`` macro to create a stand-alone linear constraint instance.
+  * ``getValue()`` method implemented to compute the value of a nonlinear subexpression
 
 Version 0.8.0 (February 17, 2015)
 ---------------------------------